perf(rwkv7): skip redundant recurrent_state.to(float32) in decode
The recurrent state is allocated as float32 by get_state_dtype,
so the defensive .to(torch.float32) in forward_decode_batch is
a no-op that still incurs dispatch overhead per decode step (61
layers x 1024 steps = 62K calls per request).

+++ vllm_ascend/models/rwkv7.py
@@ -1065,7 +1065,8 @@ class RWKV7Attention(nn.Module):
             delta,
             v_first,
         )
-        recurrent_state = recurrent_state.to(torch.float32)
+        # recurrent_state is allocated as float32 by get_state_dtype;
+        # skip the redundant .to() call to avoid dispatch overhead.
         final_recurrent_state = _rwkv7_recurrent_step(
             recurrent_state,
             w,
